pyproject.toml
[project]
name = "PipeDSL"
version = "0.1.8"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "nltk (>=3.9.1,<3.9.2)",
    "pytest (>=8.3.5,<9.0.0)",
    "jsonpath-ng (>=1.7.0,<2.0.0)",
    "coverage (>=7.10.6,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)"
]


[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
exclude = [
    "tests/*",
]
ignore_missing_imports = true


[tool.hatch.build.targets.wheel]
packages = ["PipeDSL/"]

[build-system]
requires = ["hatchling >= 1.27"]
build-backend = "hatchling.build"

